requirements.txt
Django==1.11.1
djangorestframework==3.6.3
django-oauth-toolkit==0.12.0
boto3==1.3.0